If keypoints are provided without class_ids, default to class_ids=0
